Delete existing files while using requests
I don't know how I managed to miss it for this long,
but when "Updating" with requests, if a file has been
deleted from gist but still exists in the computer,
when updating the gist with requests, it will keep it
in the folder. I just used shutil.rmtree (rm -rf basically)
to start from the beginning.